Branch params/route set

+++ app/src/main/java/com/experlabs/training/activities/MainActivity.kt
@@ -19,11 +19,14 @@ import org.koin.android.ext.android.inject
 import org.koin.android.scope.AndroidScopeComponent
 import org.koin.androidx.scope.activityScope
 import org.koin.core.scope.Scope
+import retrofit2.Response
 
 class MainActivity : AppCompatActivity(), AndroidScopeComponent {
     private lateinit var adapter: MemeMainAdapter
-    private lateinit var binding : ActivityMainBinding
-    private lateinit var memeslist : MutableList<Meme>
+    private lateinit var binding: ActivityMainBinding
+    private lateinit var memeslist: MutableList<Meme>
+    private var index = 0
+    private var branch_flag = false
 
     override val scope: Scope by activityScope()
     private val memeViewModel: MemeViewModel by inject<MemeViewModel>()
@@ -31,6 +34,12 @@ class MainActivity : AppCompatActivity(), AndroidScopeComponent {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
+        branchListener.branchResponse{flag, value ->
+            branch_flag = flag
+            index = value
+            binding.mainRecycler.scrollToPosition(index)
+            Toast.makeText(this, "Scrolled to $index", Toast.LENGTH_SHORT).show()
+        }
 
         binding = ActivityMainBinding.inflate(layoutInflater)
         val view = binding.root
@@ -43,40 +52,58 @@ class MainActivity : AppCompatActivity(), AndroidScopeComponent {
 
         binding.deleteAllBt.setOnClickListener {
             memeslist.clear()
-            memeViewModel.deleteAllMemesFromRepository()
-            adapter.notifyDataSetChanged()
-            Toast.makeText(this, "Deleted All", Toast.LENGTH_SHORT).show()
+            memeViewModel.deleteAllMemesFromRepository() { flag, message ->
+
+                if (flag)
+                    adapter.notifyDataSetChanged()
+                Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
+            }
+
         }
 
-        memeViewModel.getMemesFromRepository("", false) { _, _ -> }
-        memeViewModel.memes.observe(this, Observer<List<Meme>> { data ->
-            data?.let { memes ->
-                memeslist = memes as MutableList<Meme>
+        memeViewModel.getMemesFromRepository("", false) { flag, message ->
 
-                Toast.makeText(this, "${memes.size}", Toast.LENGTH_SHORT).show()
+            if (flag) {
+                memeViewModel.memes.observe(this, Observer<List<Meme>> { data ->
+                    data?.let { memes ->
+                        memeslist = memes as MutableList<Meme>
 
-                adapter = MemeMainAdapter(memeslist) { item, index ->
-                    deleteItem(item, index)
+                        Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
 
-                }
-                binding.mainRecycler.adapter = adapter
-                binding.deleteAllBt.isEnabled = true
+                        adapter = MemeMainAdapter(memeslist) { item, index ->
+                            deleteItem(item, index)
+
+                        }
+                        binding.mainRecycler.adapter = adapter
+
+                        binding.deleteAllBt.isEnabled = true
+                    }
+                })
+            } else {
+                binding.mainRecycler.adapter = null
+                Toast.makeText(this, message, Toast.LENGTH_SHORT).show()
             }
-        })
+
+        }
     }
 
-    private fun deleteItem(item: Meme, index : Int) {
-        memeViewModel.deleteMemeFromRepository(item)
-        memeslist.removeAt(index)
-        adapter.notifyItemRemoved(index)
-        adapter.notifyItemRangeChanged(index, 1)
-        Toast.makeText(this, "${item.name} Deleted!", Toast.LENGTH_SHORT).show()
+    private fun deleteItem(item: Meme, index: Int) {
+        memeViewModel.deleteMemeFromRepository(item) { flag, message ->
+
+            if (flag) {
+                memeslist.removeAt(index)
+                adapter.notifyItemRemoved(index)
+                adapter.notifyItemRangeChanged(index, 1)
+                Toast.makeText(this, "${item.name} : $message!", Toast.LENGTH_SHORT).show()
+            }
+        }
     }
 
     override fun onStart() {
         super.onStart()
         // Branch init
-        Branch.sessionBuilder(this).withCallback(branchListener).withData(this.intent?.data).init()
+        Branch.sessionBuilder(this).withCallback(branchListener).withData(this.intent?.data)
+            .init()
     }
 
     override fun onNewIntent(intent: Intent) {
@@ -85,23 +112,39 @@ class MainActivity : AppCompatActivity(), AndroidScopeComponent {
 
         // if activity is in foreground (or in backstack but partially visible) launch the same
         // activity will skip onStart, handle this case with reInit
-        if (intent != null &&
-            intent.hasExtra("branch_force_new_session") &&
-            intent.getBooleanExtra("branch_force_new_session", false)
-        ) {
+        if (intent.hasExtra("branch_force_new_session") && intent.getBooleanExtra("branch_force_new_session", false)) {
             Branch.sessionBuilder(this).withCallback(branchListener).reInit()
         }
     }
 
     object branchListener : Branch.BranchReferralInitListener {
+
+        public lateinit var response : (Boolean, Int) -> Unit
+
         override fun onInitFinished(referringParams: JSONObject?, error: BranchError?) {
             if (error == null) {
                 Log.i("BRANCH SDK", referringParams.toString())
-                // Retrieve deeplink keys from 'referringParams' and evaluate the values to determine where to route the user
+
+                referringParams?.let {
+                    if (it.getString("+clicked_branch_link").toBoolean()) {
+                        it.getString("index").let { value ->
+                            response(true, value.toInt())
+                            return
+                        }
+                    }
+                }
+                response(false, 0)
+
+            // Retrieve deeplink keys from 'referringParams' and evaluate the values to determine where to route the user
                 // Check '+clicked_branch_link' before deciding whether to use your Branch routing logic
             } else {
+                response(false, 0)
                 Log.e("BRANCH SDK", error.message)
             }
+        }
+
+        fun branchResponse(callback : (Boolean, Int) -> Unit) {
+            response = callback
         }
     }
 }
